requirements.txt
python-dotenv
mangum
fastapi
uvicorn
